requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pymongo>=4.6.0
python-dotenv>=1.0.0
openai>=1.54.0
pydantic>=2.10.0
python-multipart>=0.0.12

